+++ CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5)
 
 # Set extension name here
-set(TARGET_NAME quack)
+set(TARGET_NAME rate_limit_filesystem)
 project(${TARGET_NAME})
 
 set(CMAKE_CXX_STANDARD 17)
