+++ Makefile
@@ -1,7 +1,7 @@
 PROJ_DIR := $(dir $(abspath $(lastword $(MAKEFILE_LIST))))
 
 # Configuration of extension
-EXT_NAME=quack
+EXT_NAME=rate_limit_filesystem
 EXT_CONFIG=${PROJ_DIR}extension_config.cmake
 
 # Include the Makefile from extension-ci-tools
@@ -11,12 +11,12 @@ format-all: format
 	find test/unittest -iname *.hpp -o -iname *.cpp | xargs clang-format --sort-includes=0 -style=file -i
 
 test_unit: all
-	find build/release/extension/quack/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
+	find build/release/extension/rate_limit_filesystem/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
 
 test_reldebug_unit: all
-	find build/reldebug/extension/quack/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
+	find build/reldebug/extension/rate_limit_filesystem/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
 
 test_debug_unit: reldebug
-	find build/debug/extension/quack/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
+	find build/debug/extension/rate_limit_filesystem/ -type f -name "test*" -not -name "*.o" -not -name "*.cpp" -not -name "*.d" -exec {} \;
 
 PHONY: format-all test_unit test_reldebug_unit test_debug_unit
