+++ docs/NEXT_README.md
@@ -1,10 +1,10 @@
-# Quack
+# Rate Limit Filesystem
 
 This repository is based on https://github.com/duckdb/extension-template, check it out if you want to build and ship your own DuckDB extension.
 
 ---
 
-This extension, Quack, allow you to ... <extension_goal>.
+This extension, Rate Limit Filesystem, provides rate limiting capabilities for filesystem operations.
 
 
 ## Building
@@ -26,24 +26,24 @@ The main binaries that will be built are:
 ```sh
 ./build/release/duckdb
 ./build/release/test/unittest
-./build/release/extension/quack/quack.duckdb_extension
+./build/release/extension/rate_limit_filesystem/rate_limit_filesystem.duckdb_extension
 ```
 - `duckdb` is the binary for the duckdb shell with the extension code automatically loaded.
 - `unittest` is the test runner of duckdb. Again, the extension is already linked into the binary.
-- `quack.duckdb_extension` is the loadable binary as it would be distributed.
+- `rate_limit_filesystem.duckdb_extension` is the loadable binary as it would be distributed.
 
 ## Running the extension
 To run the extension code, simply start the shell with `./build/release/duckdb`.
 
-Now we can use the features from the extension directly in DuckDB. The template contains a single scalar function `quack()` that takes a string arguments and returns a string:
+Now we can use the features from the extension directly in DuckDB. The template contains a single scalar function `rate_limit_filesystem()` that takes a string arguments and returns a string:
 ```
-D select quack('Jane') as result;
-┌───────────────┐
-│    result     │
-│    varchar    │
-├───────────────┤
-│ Quack Jane 🐥 │
-└───────────────┘
+D select rate_limit_filesystem('Jane') as result;
+┌────────────────────────────────┐
+│            result              │
+│            varchar             │
+├────────────────────────────────┤
+│ RateLimitFilesystem Jane 🐥   │
+└────────────────────────────────┘
 ```
 
 ## Running the tests
@@ -81,8 +81,8 @@ DuckDB. To specify a specific version, you can pass the version instead.
 
 After running these steps, you can install and load your extension using the regular INSTALL/LOAD commands in DuckDB:
 ```sql
-INSTALL quack;
-LOAD quack;
+INSTALL rate_limit_filesystem;
+LOAD rate_limit_filesystem;
 ```
 
 ## Setting up CLion
