+++ docs/README.md
@@ -80,15 +80,15 @@ GEN=ninja make
 ## Running the extension
 To run the extension code, simply start the shell with `./build/release/duckdb`. This shell will have the extension pre-loaded.  
 
-Now we can use the features from the extension directly in DuckDB. The template contains a single scalar function `quack()` that takes a string arguments and returns a string:
+Now we can use the features from the extension directly in DuckDB. The template contains a single scalar function `rate_limit_filesystem()` that takes a string arguments and returns a string:
 ```
-D select quack('Jane') as result;
-┌───────────────┐
-│    result     │
-│    varchar    │
-├───────────────┤
-│ Quack Jane 🐥 │
-└───────────────┘
+D select rate_limit_filesystem('Jane') as result;
+┌────────────────────────────────┐
+│            result              │
+│            varchar             │
+├────────────────────────────────┤
+│ RateLimitFilesystem Jane 🐥   │
+└────────────────────────────────┘
 ```
 
 ## Running the tests
