+++ src/include/rate_limit_filesystem_extension.hpp
@@ -4,7 +4,7 @@
 
 namespace duckdb {
 
-class QuackExtension : public Extension {
+class RateLimitFilesystemExtension : public Extension {
 public:
 	void Load(ExtensionLoader &db) override;
 	std::string Name() override;
@@ -12,3 +12,4 @@ public:
 };
 
 } // namespace duckdb
+
