+++ test/sql/rate_limit_filesystem.test
@@ -1,0 +1,24 @@
+# name: test/sql/rate_limit_filesystem.test
+# description: test rate_limit_filesystem extension
+# group: [sql]
+
+# Before we load the extension, this will fail
+statement error
+SELECT rate_limit_filesystem('Sam');
+----
+Catalog Error: Scalar Function with name rate_limit_filesystem does not exist!
+
+# Require statement will ensure this test is run with this extension loaded
+require rate_limit_filesystem
+
+# Confirm the extension works
+query I
+SELECT rate_limit_filesystem('Sam');
+----
+RateLimitFilesystem Sam 🐥
+
+query I
+SELECT rate_limit_filesystem_openssl_version('Michael') ILIKE 'RateLimitFilesystem Michael, my linked OpenSSL version is OpenSSL%';
+----
+true
+
